requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=-
#]
Flask-Cors==4.0.0
boto3==1.16.51
    # via -r requirements.in
botocore==1.19.63
    # via
    #   boto3
    #   s3transfer
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.4
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
requests==2.31.0
    # via -r requirements.in
s3transfer==0.3.7
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.18
    # via
    #   botocore
    #   requests
gunicorn===19.1.1
pip==23.3.1

tensorflow-intel==0.0.1
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
Pillow==9.5.0
python-dotenv== 0.21.1
numpy
jo
opencv-python
tensorflow